C++11 RNG
Moved the random number generation from the custom RNG to the C++11
built-in Mersenne Twister. Distributions are provided for each use case
of the RNG. The seeding logic is currently possibly wrong; need to
double check that.  Initial runs suggest that the random seeding is
substantially slower; I don't know why this would be - random_device()
is only called once, to seed the Mersenne Twister. There should be no
meaningful difference in speed.