Configuration file
To improve the comprehensibility and flexibility of the model I added a
configuration file to govern the parameters, replacing my original plan
of using CLI flags (gflags is now used to specify the filename). The
config file has all the parameters of the model, save the two most
important ones: numagents and numcommodities. Because variable-length
arrays are problematic, I can't easily declare variables associated with
those and then initialize later.  Changing this requires restructuring a
fair bit of the code, particular the Dot struct. Another side effect of
moving to the file-based approach is that I have to declare and
initialize the random distributions once and then re-initialize them
with the correct parameters. I can probably declare the distributions in
file scope and then initialize later; I'll loook into this tomorrow.