BREAKING CHANGE - refactor to vectors from arrays
Currently broken, thus the branching - I am moving some key arrays to
vectors so that they can be dynamically sized at runtime, allowing
NumberOfAgents and NumberOfCommodities to be passed as runtime
parameters. I have already fixed the crucial CommodityArray typedef; I
still need to fix the Data object, which seems to require some changes
throughout the code.  Note also that this refactor has caused the
compiler to throw a number of warnings about implicit conversions - this
makes me realize that most of the ints in the model should probably
unsigned ints. This is probably not too much trouble but still needs to
be dealt with in some way, the sheer number of warnings is distracting
and I don't want to disable -Wsign-conversion.